Possibly resolve issues with opening windows during tests
